[INLONG-8154][Manager] modify the name.
